+++ pallets/astar-xcm-benchmarks/src/generic/mod.rs
@@ -18,6 +18,7 @@
 
 pub use pallet::*;
 
+#[cfg(feature = "runtime-benchmarks")]
 pub mod benchmarking;
 
 #[frame_support::pallet]
